// src/app/protected/products/catagory.interface.ts
export interface Category {
    index: number;
    categoryId: number;
    categoryCTId: string;
    categoryName: string;
    categorySeq: string;
    categoryComment: string;
    categoryLevel: number;
    categoryImage: string;
    categoryStatus: boolean;
    categoryLastupdate: any;
    categoryCreated: any;
    // index: number;
    // categoryId: number;
    // categoryCTId: string;
    // categoryName: string;
    // categoryComment: string;
    // categoryLastupdate: any;
    // categoryCreated: Date;
    // categoryImage: string;
}